+++ crates/teloxide-macros/src/command.rs
@@ -13,10 +13,14 @@ pub(crate) struct Command {
     pub description: Option<(String, bool, Span)>,
     /// Name of the command, with all renames already applied.
     pub name: String,
+    /// The aliases of the command.
+    pub aliases: Option<(Vec<String>, Span)>,
     /// Parser for arguments of this command.
     pub parser: ParserType,
     /// Whether the command is hidden from the help message.
     pub hidden: bool,
+    /// Whether the aliases of the command are hidden from the help message.
+    pub hidden_aliases: bool,
 }
 
 impl Command {
@@ -31,12 +35,14 @@ impl Command {
             description,
             rename_rule,
             rename,
+            aliases,
             parser,
             // FIXME: error on/do not ignore separator
             separator: _,
             // FIXME: error on/do not ignore command separator
             command_separator: _,
             hide,
+            hide_aliases,
         } = attrs;
 
         let name = match (rename, rename_rule) {
@@ -54,13 +60,21 @@ impl Command {
         let prefix = prefix.map(|(p, _)| p).unwrap_or_else(|| global_options.prefix.clone());
         let parser = parser.map(|(p, _)| p).unwrap_or_else(|| global_options.parser_type.clone());
         let hidden = hide.is_some();
+        let hidden_aliases = hide_aliases.is_some();
 
-        Ok(Self { prefix, description, parser, name, hidden })
+        Ok(Self { prefix, description, parser, name, aliases, hidden, hidden_aliases })
     }
 
     pub fn get_prefixed_command(&self) -> String {
         let Self { prefix, name, .. } = self;
         format!("{prefix}{name}")
+    }
+
+    pub fn get_prefixed_aliases(&self) -> Option<Vec<String>> {
+        let Self { prefix, aliases, .. } = self;
+        aliases
+            .as_ref()
+            .map(|(aliases, _)| aliases.iter().map(|alias| format!("{prefix}{alias}")).collect())
     }
 
     pub fn description(&self) -> Option<&str> {